+++ src/components/CocktailCard.js
@@ -1,17 +1,17 @@
-import React from "react";
-import { updateFavourites } from "../actions";
-import { connect } from "react-redux";
-import { isFavouriteSelector, allGlassesSelector } from "../selectors";
+import React, { useState } from "react";
+import { connect, useSelector } from "react-redux";
+import { useTheme } from "@mui/material/styles";
+import { allGlassesSelector } from "../selectors";
 import CocktailAvatar from "./CocktailAvatar";
 import {
   Card,
   CardHeader,
   CardContent,
-  CardActions,
   CardActionArea,
+  CardActions,
   Typography,
+  Button,
 } from "@mui/material";
-import CocktailActions from "./CocktailActions";
 import GlassIcon from "./GlassIcon";
 import VeganIcon from "@mui/icons-material/FilterVintage";
 
@@ -19,15 +19,35 @@ import Redo from "@mui/icons-material/Redo";
 
 import Ingredient from "./IngredientDetail";
 import { Link } from "react-router-dom";
-import { bindActionCreators } from "redux";
+import DispenseWorkflow from "./CocktailPage/DispenseWorkflow";
 
-const CocktailCard = ({
-  cocktail,
-  allGlasses,
-  favourite,
-  favourites,
-  updateFavourites,
-}) => {
+const CocktailCard = ({ cocktail, allGlasses }) => {
+  const theme = useTheme();
+  const [showDispense, setShowDispense] = useState(false);
+
+  const robotUrl = useSelector(
+    (state) => state.settings.robot && state.settings.robot.url,
+  );
+  const robotConnected = useSelector((state) => state.robot.connected);
+  const robotState = useSelector((state) => state.robot.robotState);
+  const bar = useSelector((state) => state.bar);
+
+  const robotIdle = robotConnected && robotState && robotState.state === "idle";
+
+  const hasDispensable =
+    cocktail &&
+    cocktail.ingredients.some((ing) =>
+      bar.some(
+        (item) =>
+          item &&
+          (ing.ingredient === item.type || ing.ingredient === item.ingredient),
+      ),
+    );
+
+  const canDispense = !!robotUrl && robotIdle && hasDispensable;
+
+  const mixButtonLabel = theme.custom?.mixButtonLabel ?? "Mix it!";
+
   if (!cocktail) return null;
 
   return (
@@ -100,22 +120,30 @@ const CocktailCard = ({
           )}
         </CardContent>
       </CardActionArea>
-
-      <CardActions sx={{ alignSelf: "flex-end", flexGrow: 1 }}>
-        <CocktailActions cocktail={cocktail} />
-      </CardActions>
+      {canDispense && (
+        <CardActions>
+          <Button
+            size="small"
+            variant="contained"
+            color="primary"
+            onClick={() => setShowDispense(true)}
+          >
+            {mixButtonLabel}
+          </Button>
+        </CardActions>
+      )}
+      {showDispense && (
+        <DispenseWorkflow
+          cocktail={cocktail}
+          onClose={() => setShowDispense(false)}
+        />
+      )}
     </Card>
   );
 };
 
-const mapStateToProps = (state, ownProps) => ({
-  favourite: isFavouriteSelector(state, ownProps),
-  favourites: state.favourites,
+const mapStateToProps = (state) => ({
   allGlasses: allGlassesSelector(state),
 });
 
-const mapDispatchToProps = (dispatch) => ({
-  updateFavourites: bindActionCreators(updateFavourites, dispatch),
-});
-
-export default connect(mapStateToProps, mapDispatchToProps)(CocktailCard);
+export default connect(mapStateToProps)(CocktailCard);