+++ src/themes/c64/theme.js
@@ -74,5 +74,6 @@ export default createTheme({
   },
   custom: {
     cardWidth: 640,
+    mixButtonLabel: "RUN",
   },
 });